// examples/nft/components/types.tsx
import {
  ConfigResponse as VendingMinterConfigResponse,
  MintCountResponse,
  MintableNumTokensResponse,
  MintPriceResponse,
} from 'stargazejs/types/codegen/VendingMinter.types';
import { ConfigResponse as WhitelistConfigResponse } from 'stargazejs/types/codegen/Whitelist.types';
import {
  ContractInfoResponse,
  CollectionInfoResponse,
} from 'stargazejs/types/codegen/SG721Base.types';

export enum TxResult {
  Success = 0,
  Failed = 1,
}

export enum SaleType {
  FIXED_PRICE = 'fixed_price',
  AUCTION = 'auction',
}

export type Rename<T, K extends keyof T, R extends PropertyKey> = Omit<T, K> & {
  [P in R]: T[K];
};

export type Minter = VendingMinterConfigResponse &
  Rename<MintableNumTokensResponse, 'count', 'remaining_tokens'> &
  Rename<Omit<MintCountResponse, 'address'>, 'count', 'user_minted'> & {
    all_prices: MintPriceResponse;
  };

export type SG721 = CollectionInfoResponse & ContractInfoResponse;

export type Whitelist = WhitelistConfigResponse;

export type ContractsAddress = {
  minter: string;
  sg721: string;
};

export type TokensMint = {
  tokens: {
    tokens: {
      tokenId: string;
      imageUrl: string;
    }[];
  };
};

export type CollectionsMint = {
  collections: {
    collections: {
      collectionAddr: string;
      floorPrice: number;
    }[];
  };
};

export type CollectionMint = {
  collection: {
    image: string;
  };
};

// *Collection Types*
export interface Collections {
  collections: {
    collections: Collection[];
    total: number;
  };
}

export interface Collection {
  collectionAddr: string;
  name: string;
  description: string;
  tokensCount: number;
  createdBy: CreatedBy;
  floorPrice: number;
}

interface CreatedBy {
  addr: string;
  name?: {
    name: string;
  };
}

// *Token Types*
export interface Tokens {
  tokens: {
    total: number;
    tokens: Token[];
  };
}

export interface Token {
  collectionAddr: string;
  name: string;
  price?: Price;
  priceExpiresAt?: string;
  saleType?: string;
  tokenId: string;
  highestCollectionBid: string;
  description: string;
  forSale: boolean;
  imageUrl: string;
  rarityOrder: number;
  traits: Trait[];
  id: string;
  lastSale: LastSale;
  highestCollectionBidEvent: {
    bidder: {
      addr: string;
    };
  };
}

interface LastSale {
  eventName: string;
  price: Price;
}

export interface Trait {
  name: string;
  value: string;
  rarityPercent: number;
}

interface Price {
  denom: string;
  amount: string;
}

// *Name Types*
export interface Names {
  names: {
    names: Name[];
  };
}

interface Name {
  name: string;
  records: Record[];
}

interface Record {
  id: string;
  recordName: string;
  recordValue: string;
  verified: boolean;
  createdAt: string;
  updatedAt: string;
  name: string;
  contractAddr: string;
}

// *Event Types*
export interface Events {
  events: {
    edges: Edge[];
  };
}

interface Edge {
  node: Node;
}

interface Node {
  eventName: string;
  price?: Price;
  data: Data;
  action?: string;
  id: string;
}

interface Data {
  seller?: string;
  collection?: string;
  starsUsdRatio?: string;
  buyer?: string;
  price?: string;
  tokenId: string;
  priceUsd?: string;
  saleType?: string;
  expires?: string;
  bidPrice?: string;
  bidPriceUsd?: string;
  bidder?: string;
  recipient?: string;
  sender?: string;
  spender?: string;
  minter?: string;
  networkFee?: string;
  sellerAmount?: string;
  mintPrice?: string;
}

// *Contract Types*

export interface Contract {
  contract: {
    contractInfo: ContractInfo;
  };
}

interface ContractInfo {
  start_trading_time: string;
  royalty_info: Royaltyinfo;
  description: string;
  image: string;
  creator: string;
  explicit_content: boolean;
  external_link?: any;
}

interface Royaltyinfo {
  payment_address: string;
  share: string;
}